This is mainly a code refactoring to split code from app file into smaller parts by moving code into screens and widgets. It also fixes #7 and #10
