+++ tygenie/screen.py
@@ -1,0 +1,26 @@
+from textual.screen import Screen
+
+
+class TyScreen(Screen):
+
+    def bind(
+        self,
+        keys: str,
+        action: str,
+        *,
+        description: str = "",
+        show: bool = True,
+        key_display: str | None = None,
+    ) -> None:
+        """Bind a key to an action.
+
+        Args:
+            keys: A comma separated list of keys, i.e.
+            action: Action to bind to.
+            description: Short description of action.
+            show: Show key in UI.
+            key_display: Replacement text for key, or None to use default.
+        """
+        self._bindings.bind(
+            keys, action, description, show=show, key_display=key_display
+        )
